requirements.txt
mysql-connector-python==8.0.33
pydantic==1.10.8
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.15